Created server class with seperate cpp/hpp files
